+++ packages/node/src/codecs.ts
@@ -1,25 +1,45 @@
-export { imageEncode, imageDecode };
+export { imageEncode, imageDecode, MimeType };
 
 import sharp, { FormatEnum } from 'sharp';
 import ndarray, { NdArray } from 'ndarray';
 
 async function imageDecode(blob: Blob): Promise<NdArray<Uint8Array>> {
   const buffer = await blob.arrayBuffer();
-  const decoded = sharp(buffer);
 
-  let { width, height, channels } = await decoded.metadata();
+  const mime = MimeType.fromString(blob.type);
+  switch (mime.type) {
+    case 'image/x-alpha8': {
+      const width = parseInt(mime.params['width']);
+      const height = parseInt(mime.params['height']);
+      return ndarray(await blob.arrayBuffer(), [height, width, 1]);
+    }
+    case 'image/x-rgba8': {
+      const width = parseInt(mime.params['width']);
+      const height = parseInt(mime.params['height']);
+      return ndarray(await blob.arrayBuffer(), [height, width, 4]);
+    }
+    case 'application/octet-stream':
+    case `image/png`:
+    case `image/jpeg`:
+    case `image/webp`: {
+      const decoded = sharp(buffer);
+      let { width, height, channels } = await decoded.metadata();
 
-  if (channels === 3) {
-    decoded.ensureAlpha();
-    channels = 4;
+      if (channels === 3) {
+        decoded.ensureAlpha();
+        channels = 4;
+      }
+
+      const outBuffer = await decoded.raw().toBuffer();
+      const array = ndarray(outBuffer, [height!, width!, channels!]);
+      await sharp(array.data, {
+        raw: { width: width!, height: height!, channels: channels! }
+      });
+      return array;
+    }
+    default:
+      throw new Error(`Unsupported format: ${mime.type}`);
   }
-
-  const outBuffer = await decoded.raw().toBuffer();
-  const array = ndarray(outBuffer, [height!, width!, channels!]);
-  await sharp(array.data, {
-    raw: { width: width!, height: height!, channels: channels! }
-  });
-  return array;
 }
 
 async function imageEncode(
@@ -33,10 +53,11 @@ async function imageEncode(
   type Keys = keyof FormatEnum;
 
   switch (type) {
-    case 'image/x-r8':
-      return new Blob([imageTensor.data], { type: 'image/x-alpha8' });
-    case 'image/x-rgba8':
-      return new Blob([imageTensor.data], { type: 'image/x-rgba8' });
+    case 'image/x-alpha8':
+    case 'image/x-rgba8': {
+      const mime = MimeType.create(type, { width, height });
+      return new Blob([imageTensor.data], { type: mime.toString() });
+    }
     case `image/png`:
     case `image/jpeg`:
     case `image/webp`:
@@ -47,5 +68,47 @@ async function imageEncode(
       return new Blob([buffer], { type: type });
     default:
       throw new Error(`Invalid format: ${format}`);
+  }
+}
+
+class MimeType {
+  type: string = 'application/octet-stream';
+  params: Record<string, string> = {};
+
+  private constructor(type: string, params: Record<string, string>) {
+    this.type = type;
+    this.params = params;
+  }
+
+  toString(): string {
+    const paramsStr = [];
+    for (const key in this.params) {
+      const value = this.params[key];
+      paramsStr.push(`${key}=${value}`);
+    }
+    return [this.type, ...paramsStr].join(';');
+  }
+
+  static create(type, params: Record<string, string>): MimeType {
+    return new MimeType(type, params);
+  }
+
+  isIdentical(other: MimeType): Boolean {
+    return this.type === other.type && this.params === other.params;
+  }
+
+  isEqual(other: MimeType): Boolean {
+    return this.type === other.type;
+  }
+
+  static fromString(mimeType: string): MimeType {
+    const [type, ...paramsArr] = mimeType.split(';');
+    const params: Record<string, string> = {};
+
+    for (const param of paramsArr) {
+      const [key, value] = param.split('=');
+      params[key.trim()] = value.trim();
+    }
+    return new MimeType(type, params);
   }
 }