+++ packages/web/src/index.ts
@@ -1,5 +1,11 @@
 export default removeBackground;
-export { preload, removeBackground, removeForeground, segmentForeground };
+export {
+  preload,
+  removeBackground,
+  removeForeground,
+  segmentForeground,
+  applySegmentationMask
+};
 export type { Config, ImageSource };
 
 import { memoize } from 'lodash';
@@ -100,35 +106,69 @@ async function segmentForeground(
   const [height, width, channels] = imageTensor.shape;
 
   const alphamask = await runInference(imageTensor, config, session);
-  const stride = width * height;
 
-  if (config.output.format === 'image/x-alpha8') {
-    const outImage = await utils.imageEncode(
-      alphamask,
-      config.output.quality,
-      config.output.format
-    );
-    return outImage;
-  } else {
-    const outImageTensor = ndarray(new Uint8Array(channels * stride), [
-      height,
-      width,
-      channels
-    ]);
-    for (let i = 0; i < stride; i += 1) {
-      const index = 4 * i + 3;
-      outImageTensor.data[index] = alphamask.data[i]; //Red
-      outImageTensor.data[index + 1] = alphamask.data[i]; //Green
-      outImageTensor.data[index + 2] = alphamask.data[i]; // Blue
-      outImageTensor.data[index + 3] = 255;
+  switch (config.output.format) {
+    case 'image/x-alpha8': {
+      const outImage = await utils.imageEncode(
+        alphamask,
+        config.output.quality,
+        config.output.format
+      );
+      return outImage;
     }
-
-    const outImage = await utils.imageEncode(
-      outImageTensor,
-      config.output.quality,
-      config.output.format
-    );
-
-    return outImage;
+    default: {
+      const stride = width * height;
+      const outImageTensor = ndarray(new Uint8Array(channels * stride), [
+        height,
+        width,
+        channels
+      ]);
+      for (let i = 0; i < stride; i += 1) {
+        const index = 4 * i + 3;
+        outImageTensor.data[index] = alphamask.data[i]; //Red
+        outImageTensor.data[index + 1] = alphamask.data[i]; //Green
+        outImageTensor.data[index + 2] = alphamask.data[i]; // Blue
+        outImageTensor.data[index + 3] = 255;
+      }
+      const outImage = await utils.imageEncode(
+        outImageTensor,
+        config.output.quality,
+        config.output.format
+      );
+      return outImage;
+    }
   }
+}
+
+async function applySegmentationMask(
+  image,
+  mask,
+  config?: Config
+): Promise<Blob> {
+  config = validateConfig(config);
+  const imageTensor = await utils.imageSourceToImageData(image, config);
+  const [imageHeight, imageWidth, imageChannels] = imageTensor.shape;
+  const maskTensor = await utils.imageSourceToImageData(mask, config);
+  const [maskHeight, maskWidth, maskChannels] = maskTensor.shape;
+  console.log([maskHeight, maskWidth, maskChannels]);
+  console.log([imageHeight, imageWidth, imageChannels]);
+
+  const alphaMask =
+    maskHeight !== imageHeight || maskWidth !== imageWidth
+      ? utils.tensorResizeBilinear(maskTensor, imageWidth, imageHeight)
+      : maskTensor;
+  const stride = imageWidth * imageHeight;
+  for (let i = 0; i < stride; i += 1) {
+    const idxImage = imageChannels * i;
+    const idxMask = maskChannels * i;
+    imageTensor.data[idxImage + 3] = alphaMask.data[idxMask]; // alpha information it always in the first (sometimes also in the others)
+  }
+
+  const outImage = await utils.imageEncode(
+    imageTensor,
+    config.output.quality,
+    config.output.format
+  );
+
+  return outImage;
 }